Poprawa typu zmiennej w klasie ApartmentSerrlement

+++ main.py
@@ -20,7 +20,7 @@ class ApartmentSettlement(BaseModel):
     month: int 
     year: int 
     total_bills: int 
-    sum_of_rent: int
+    sum_of_rent: float
 
 
 class Apartment(BaseModel):
